Fix UI after review
